requirements.txt
alembic==1.8.0
click==8.1.3
Flask==2.1.2
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
ifcfg==0.22
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
python-dotenv==0.20.0
SQLAlchemy==1.4.37
Werkzeug==2.1.2
WTForms==3.0.1
